+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugfixes
 - [#2084](https://github.com/influxdb/influxdb/pull/2084): Allowing leading underscores in identifiers.
 - [#2080](https://github.com/influxdb/influxdb/pull/2080): Graphite logs in seconds, not milliseconds.
+- [#2101](https://github.com/influxdb/influxdb/pull/2101): SHOW DATABASES should name returned series "databases".
 
 ## v0.9.0-rc16 [2015-03-24]